Fixed bug preventing tableView from being populated with object data.
