+++ Journal Core Data/Journal Core Data/Models/Entry+Convience.swift
@@ -1,0 +1,20 @@
+//
+//  Entry+Convience.swift
+//  Journal Core Data
+//
+//  Created by Austin Cole on 1/14/19.
+//  Copyright © 2019 Austin Cole. All rights reserved.
+//
+
+import Foundation
+import CoreData
+
+extension Entry {
+    convenience init(title: String, bodyText: String, timestamp: Date = Date.init(), identifier: String = UUID().uuidString, insertInto context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
+        self.init(context: context)
+        self.title = title
+        self.bodyText = bodyText
+        self.timestamp = timestamp
+        self.identifier = identifier
+    }
+}
