--- Journal Core Data/Journal Core Data/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Journal Core Data
-//
-//  Created by Austin Cole on 1/14/19.
-//  Copyright © 2019 Austin Cole. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-
-}
-
