+++ Journal Core Data/Journal Core Data/Views/EntryTableViewCell.swift
